perf(frontend): merge SearchDialog dead-link and action variants
Parameterize dead-wiki-link rewrites; hoist search helpers; mark three
file-family slices done in the frontend test-opt plan.
